class 043 (07-14)

+++ 037-laravel/first-laravel/routes/web.php
@@ -38,36 +38,36 @@ Route::post('/skirtumas', [sum::class, 'calculate']) -> name('skaiciuotuvas');
 
 // COLORS ROUTERS
 
-Route::get('/colors', [C::class, 'index'])->name('color-index');
+Route::get('/colors', [C::class, 'index'])->name('color-index')->middleware('rw:user');
 
-Route::get('/colors/create', [C::class, 'create'])->name('color-create');
+Route::get('/colors/create', [C::class, 'create'])->name('color-create')->middleware('rw:admin');
 
-Route::post('/colors', [C::class, 'store'])->name('color-store');
+Route::post('/colors', [C::class, 'store'])->name('color-store')->middleware('rw:admin');
 
-Route::get('/colors/edit/{color}', [C::class, 'edit'])->name('color-edit');
+Route::get('/colors/edit/{color}', [C::class, 'edit'])->name('color-edit')->middleware('rw:admin');
 
-Route::put('/colors/{color}', [C::class, 'update'])->name('color-update');
+Route::put('/colors/{color}', [C::class, 'update'])->name('color-update')->middleware('rw:admin');
 
-Route::delete('/colors/{color}', [C::class, 'destroy'])->name('color-delete');
+Route::delete('/colors/{color}', [C::class, 'destroy'])->name('color-delete')->middleware('rw:admin');
 
-Route::get('/colors/show/{id}', [C::class, 'show'])->name('color-show');
+Route::get('/colors/show/{id}', [C::class, 'show'])->name('color-show')->middleware('rw:user');
 
 
 // ANIMAL ROUTERS
 
-Route::get('/animals', [A::class, 'index'])->name('animal-index');
+Route::get('/animals', [A::class, 'index'])->name('animal-index')->middleware('rw:user');
 
-Route::get('/animals/create', [A::class, 'create'])->name('animal-create');
+Route::get('/animals/create', [A::class, 'create'])->name('animal-create')->middleware('rw:admin');
 
-Route::post('/animals', [A::class, 'store'])->name('animal-store');
+Route::post('/animals', [A::class, 'store'])->name('animal-store')->middleware('rw:admin');
 
-Route::get('/animals/edit/{animal}', [A::class, 'edit'])->name('animal-edit');
+Route::get('/animals/edit/{animal}', [A::class, 'edit'])->name('animal-edit')->middleware('rw:admin');
 
-Route::put('/animals/{animal}', [A::class, 'update'])->name('animal-update');
+Route::put('/animals/{animal}', [A::class, 'update'])->name('animal-update')->middleware('rw:admin');
 
-Route::delete('/animals/{animal}', [A::class, 'destroy'])->name('animal-delete');
+Route::delete('/animals/{animal}', [A::class, 'destroy'])->name('animal-delete')->middleware('rw:admin');
 
-Route::get('/animals/show/{id}', [A::class, 'show'])->name('animal-show');
+Route::get('/animals/show/{id}', [A::class, 'show'])->name('animal-show')->middleware('rw:user');
 
 Auth::routes();
